Add players to game graphql representation
This requires splitting apart DB models and graphql models. The DB
models can only be instantiated with DB columns as fields due to Diesel
limitations. The graphql model for a game needs to be able to query the
DB to identify the game's players. However we can't instantiate a DB
game model with a connection to the DB, so we need to handle that
game->player mapping in a graphql game model instead